requirements.txt
fastapi
uvicorn
langchain==0.1.0
langchain-core==0.1.10
langchain-openai
python-dotenv
pydantic